Move and rename FlatpakPayload
Move FlatpakPayload to the RPM OSTree payload module and rename to
FlatpakManager.

Rename also the is_available method to is_source_available.

Adjust the rest of code and tests accordingly.
